Adding a products by taxon query
This PR adds a new method to the Spree Client, which allows us to fetch all the products in a given taxonomy. This will be used to power the On Demand product list. Note that for now, this just returns whatever products the Spree backend believes have the *exact* taxon specified, not any children of that taxon.

This PR also clarifies the necessary Ruby version for this repo, which was a bit of trial and error to discover.

[OD-9](https://jira.blueapron.com/browse/OD-9)