Add support for passing cart_type to Marketplace
This PR adds support for requesting orders with a specific cart_type when asking for the current order. This will allow us to filter out on-demand orders historically, and not mixing our carts.

Note: this depends on https://github.com/blueapron/spree-client-ruby/pull/11, which should go in first. (The gem file changes here are the same as that PR, modulo the gem itself)

[OD-41](https://jira.blueapron.com/browse/OD-41)